test(viam): create mock_viam_client fixture for reusable mock
